requirements.txt
-f /usr/share/pip-wheels
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
pyTelegramBotAPI==3.6.7
pytz==2019.3
requests==2.7.0
six==1.14.0
sqlparse==0.3.1
